+++ v2/connection/call.go
@@ -1,7 +1,7 @@
 //
 // DISCLAIMER
 //
-// Copyright 2020-2021 ArangoDB GmbH, Cologne, Germany
+// Copyright 2020-2024 ArangoDB GmbH, Cologne, Germany
 //
 // Licensed under the Apache License, Version 2.0 (the "License");
 // you may not use this file except in compliance with the License.
@@ -16,9 +16,6 @@
 // limitations under the License.
 //
 // Copyright holder is ArangoDB GmbH, Cologne, Germany
-//
-// Author Adam Janikowski
-// Author Tomasz Mielech
 //
 
 package connection
@@ -43,6 +40,7 @@ func CallWithChecks(ctx context.Context, c Connection, method, url string, outpu
 	}
 
 	modifiers = append(modifiers, applyGlobalSettings(ctx))
+	modifiers = append(modifiers, applyArangoDBConfiguration(c.GetConfiguration(), ctx))
 
 	for _, modifier := range modifiers {
 		if err = modifier(req); err != nil {
@@ -63,6 +61,7 @@ func CallStream(ctx context.Context, c Connection, method, url string, modifiers
 	}
 
 	modifiers = append(modifiers, applyGlobalSettings(ctx))
+	modifiers = append(modifiers, applyArangoDBConfiguration(c.GetConfiguration(), ctx))
 
 	for _, modifier := range modifiers {
 		if err = modifier(req); err != nil {