+++ v2/connection/connection_http.go
@@ -1,7 +1,7 @@
 //
 // DISCLAIMER
 //
-// Copyright 2020 ArangoDB GmbH, Cologne, Germany
+// Copyright 2020-2024 ArangoDB GmbH, Cologne, Germany
 //
 // Licensed under the Apache License, Version 2.0 (the "License");
 // you may not use this file except in compliance with the License.
@@ -17,8 +17,6 @@
 //
 // Copyright holder is ArangoDB GmbH, Cologne, Germany
 //
-// Author Adam Janikowski
-//
 
 package connection
 
@@ -29,6 +27,8 @@ type HttpConfiguration struct {
 	Endpoint       Endpoint
 
 	ContentType string
+
+	ArangoDBConfig ArangoDBConfiguration
 
 	Transport http.RoundTripper
 }
@@ -52,7 +52,7 @@ func (h HttpConfiguration) GetContentType() string {
 }
 
 func NewHttpConnection(config HttpConfiguration) Connection {
-	c := newHttpConnection(config.getTransport(), config.ContentType, config.Endpoint)
+	c := newHttpConnection(config.getTransport(), config.ContentType, config.Endpoint, config.ArangoDBConfig)
 
 	if a := config.Authentication; a != nil {
 		c.authentication = a