+++ v2/CHANGELOG.md
@@ -11,6 +11,7 @@
 - `MDI` and `MDI-Prefixed` indexes. Deprecate `ZKD` index
 - Fix url encoding for names with slashes
 - Users API support
+- Add ArangoDBConfiguration to Client config. Deprecate Context config options
 
 
 ## [2.0.3](https://github.com/arangodb/go-driver/tree/v2.0.3) (2023-10-31)